% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston}
\alias{boston}
\title{A small sample of OSM data from Boston, Massachusetts}
\format{An object of class \code{osmar} (inherits from \code{list}) of dimension 10427 x 1173 x 69.}
\source{
https://www.openstreetmap.org
}
\usage{
boston
}
\description{
A small sample of OSM data from Boston, Massachusetts
}
\keyword{datasets}
