% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{bridge_filters}
\alias{bridge_filters}
\alias{bridge_filter}
\alias{all_bridges}
\alias{main_bridges}
\title{OSM Bridge filters}
\usage{
all_bridges(graph)

main_bridges(graph)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}} object.}
}
\description{
Preset filtering functions to define which Ways in the OSM data will be
counted as bridges to be crossed.
}
\section{Functions}{
\itemize{
\item \code{all_bridges}: Include all OSM bridges

\item \code{main_bridges}: Only use major bridges, excluding any on/off ramps
}}

