% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{collect_edge_bundles}
\alias{collect_edge_bundles}
\title{Get a list of edge lists that represent multi-edge bridges}
\usage{
collect_edge_bundles(graph)
}
\arguments{
\item{A}{\code{\link{konigsberg_graph}} object}
}
\value{
A list of integer vectors representing edge indices
}
\description{
Get a list of edge lists that represent multi-edge bridges
}
