% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{get_osm_bbox}
\alias{get_osm_bbox}
\title{Download OSM data from a specified lon/lat bounding box}
\usage{
get_osm_bbox(xmin, xmax, ymin, ymax, server = "https://overpass-api.de")
}
\arguments{
\item{xmin}{Numeric. Minimum longitude}

\item{xmax}{Numeric. Maximum longitude}

\item{ymin}{Numeric. Minimum latitude}

\item{ymax}{Numeric. Maximum latitude}

\item{server}{Character. Hostname of an \href{http://overpass-api.de}{Overpass API server} Defaults to \url{https://overpass-api.de} although you may specify a different hostname.}
}
\value{
An \code{\link{osmar}} object.
}
\description{
Downloads OSM data from an \href{http://overpass-api.de}{Overpass API server} and reads it into an \code{\link{osmar}} object.
}
\seealso{
\code{\link[osmar:osmsource_api]{osmar::osmsource_api}} which iplements the same function but has stricter controls on the size that can be downloaded.
}
