% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{graph_to_sf}
\alias{graph_to_sf}
\title{Convert an igraph object with lon/lat attributes to an SF linestring collection}
\usage{
graph_to_sf(graph, v_lat, v_lon)
}
\arguments{
\item{v_lat}{Numeric. Latitude values for each vertex}

\item{v_lon}{Numeric. Longitude values for each vertex}

\item{[`igraph::igraph`]}{object}
}
\value{
A named list with an \code{edges} and a \code{nodes} \code{\link[sf:sfc]{sf::sfc}}
}
\description{
Convert an igraph object with lon/lat attributes to an SF linestring collection
}
