% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{konigsberg_graph}
\alias{konigsberg_graph}
\title{Create a road and bridge network from OSM data}
\usage{
konigsberg_graph(src, path_filter = automobile_highways,
  bridge_filter = all_bridges)
}
\arguments{
\item{src}{An \code{\link[osmar:osmar]{osmar::osmar}} object}

\item{path_filter}{A function that filters which Ways will be traversable in the graph. See \code{\link{way_filters}}.}

\item{bridge_filter}{A function that marks which Ways are bridges that need to be crossed. See \code{\link{bridge_filters}}.}
}
\value{
A \code{\link{konigsberg_graph}} object.
}
\description{
Creates a base graph object with the appropriate edge and vertex attributes from OSM.
}
