% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konigsbergr-package.R
\docType{package}
\name{konigsbergr-package}
\alias{konigsbergr}
\alias{konigsbergr-package}
\title{Find Paths Through Bridges On A Map}
\description{
A playful attempt to recreate Euler's "Bridges of
    Königsberg" problem, finding a path that crosses all the bridges in a
    city while trying to cross each bridge only once.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmu-lib/konigsbergr}
  \item Report bugs at \url{https://github.com/cmu-lib/konigsbergr/issues}
}

}
\author{
\strong{Maintainer}: Matthew Lincoln \email{mlincoln@andrew.cmu.edu} (0000-0002-4387-3384)

}
\keyword{internal}
