% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{plot_konigsberg_path}
\alias{plot_konigsberg_path}
\title{Plot a static visual of the Konigsberg pathway}
\usage{
plot_konigsberg_path(graph, pathway)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}}}

\item{pathway}{A \code{konigsberg_pathway} resulting from \code{\link[=cross_all_bridges]{cross_all_bridges()}}}
}
\value{
An \code{sf} plot
}
\description{
Plot a static visual of the Konigsberg pathway
}
