% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{select_main_component}
\alias{select_main_component}
\title{Keep only the biggest connected component of a graph}
\usage{
select_main_component(graph)
}
\arguments{
\item{graph}{A \code{\link[tidygraph:tbl_graph]{tbl_graph}} object}
}
\value{
A \code{\link[tidygraph:tbl_graph]{tbl_graph}} with one component
}
\description{
Keep only the biggest connected component of a graph
}
