% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{to_konigsberg_graph}
\alias{to_konigsberg_graph}
\title{Create a road and bridge network from OSM data}
\usage{
to_konigsberg_graph(src, path_filter = automobile_highways,
  bridge_filter = all_bridges)
}
\description{
Creates a base graph object with the appropriate edge and vertex attributes from OSM.
}
