% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{traverse_graph}
\alias{traverse_graph}
\alias{cross_all_bridges}
\alias{cross_specific_bridges}
\title{Calculate a pathways crossing bridges in konigsberg_graphs}
\usage{
cross_all_bridges(graph, starting_node = NULL, ...)

cross_specific_bridges(graph, starting_node = NULL,
  required_bridges = NULL, ...)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}}}

\item{starting_node}{An integer specifying the OSM id of the starting node. Defaults to the first vertex in \code{graph}.}

\item{...}{Additional variables passed to \code{\link[pathfinder:greedy_search]{greedy_search()}}}

\item{required_bridges}{The \href{https://wiki.openstreetmap.org/wiki/Way}{Way} or
\href{https://wiki.openstreetmap.org/wiki/Relation}{Relation} ids of bridges
that must be crossed.}
}
\value{
A \code{konigsberg_path} object, which inherits from the \code{pathfinder_path} object from \code{\link[pathfinder:greedy_search]{greedy_search()}}
}
\description{
Calculate a pathways crossing bridges in konigsberg_graphs
}
\section{Functions}{
\itemize{
\item \code{cross_all_bridges}: Cross every bridge in the graph

\item \code{cross_specific_bridges}: Cross only specified bridges
}}

