% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{view_konigsberg_path}
\alias{view_konigsberg_path}
\title{Plot bridge crossing pathway on a Leaflet map}
\usage{
view_konigsberg_path(graph, pathway)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}}}

\item{pathway}{A \code{konigsberg_pathway} resulting from \code{\link[=cross_all_bridges]{cross_all_bridges()}}}
}
\value{
A \code{\link[leaflet:leaflet]{leaflet}} object
}
\description{
This transforms the resulting pathway into an sf project and plots it onto a Leaflet map.
}
