% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{way_filters}
\alias{way_filters}
\alias{automobile_highways}
\alias{pedestrian_highways}
\title{OSM Way filters}
\usage{
automobile_highways(graph)

pedestrian_highways(graph)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}} object.}
}
\value{
An integer vector of OSM Way ids.
}
\description{
A variety of presets for filtering down a full OSM dataset to the ways that
describe transportation. This primarily filters on the value of the
\href{https://wiki.openstreetmap.org/wiki/Key:highway}{highway} tag's value.
}
\section{Functions}{
\itemize{
\item \code{automobile_highways}: Ways that are navigable in an automobile

\item \code{pedestrian_highways}: Ways that are navigable by foot
}}

\seealso{
When using \code{\link[=get_osm_file]{get_osm_file()}}, note that the underlying
\link[bigosm:read_big_osm]{read_big_osm()} function automatically subsets the
network to Ways that contain a \code{highway} key.
}
