% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{weight_by_distance}
\alias{weight_by_distance}
\title{Weights by Haversine distance between nodes}
\usage{
weight_by_distance(graph)
}
\description{
Weights by Haversine distance between nodes
}
