% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{cross_all_bridges}
\alias{cross_all_bridges}
\title{Calculate a pathway that will cross every bridge}
\usage{
cross_all_bridges(graph, starting_node = NULL, ...)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}}}

\item{starting_node}{An integer specifying the OSM id of the starting node. Defaults to the first vertex in \code{graph}.}

\item{...}{Additional variables passed to \code{\link[pathfinder:greedy_search]{pathfinder::greedy_search}}}
}
\value{
A \code{pathfinder_path} object
}
\description{
Calculate a pathway that will cross every bridge
}
