% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{get_osm_bbox}
\alias{get_osm_bbox}
\alias{get_osm_file}
\title{Download OSM data from a specified lon/lat bounding box}
\usage{
get_osm_bbox(xmin, xmax, ymin, ymax, server = "https://overpass-api.de")

get_osm_file(file)
}
\arguments{
\item{xmin}{Numeric. Minimum longitude}

\item{xmax}{Numeric. Maximum longitude}

\item{ymin}{Numeric. Minimum latitude}

\item{ymax}{Numeric. Maximum latitude}

\item{server}{Character. Hostname of an \href{http://overpass-api.de}{Overpass API server} Defaults to \url{https://overpass-api.de} although you may specify a different hostname.}

\item{file}{Filepath for an OSM XML file}
}
\value{
An \code{\link[osmar:osmar]{osmar}} object.
}
\description{
Downloads OSM data from an \href{http://overpass-api.de}{Overpass API server} and reads it into an \code{\link[osmar:osmar]{osmar}} object.
}
\section{Functions}{
\itemize{
\item \code{get_osm_file}: Read in OSM data from an XML file
}}

\seealso{
\code{\link[osmar:osmsource_api]{osmsource_api()}} which implements the same function but has stricter controls on the size that can be downloaded.
}
