% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{graph_to_sf}
\alias{graph_to_sf}
\alias{edges_to_sf}
\alias{nodes_to_sf}
\title{Convert an igraph object with lon/lat attributes to an SF linestring collection}
\usage{
graph_to_sf(graph, v_lat = igraph::vertex_attr(graph, "lat"),
  v_lon = igraph::vertex_attr(graph, "lon"))

edges_to_sf(graph, v_lat = igraph::vertex_attr(graph, "lat"),
  v_lon = igraph::vertex_attr(graph, "lon"))

nodes_to_sf(graph, v_lat = igraph::vertex_attr(graph, "lat"),
  v_lon = igraph::vertex_attr(graph, "lon"))
}
\arguments{
\item{graph}{An \code{\link[igraph:igraph]{igraph}} object}

\item{v_lat}{Numeric. Latitude values for each vertex}

\item{v_lon}{Numeric. Longitude values for each vertex}
}
\value{
A named list with an \code{edges} and a \code{nodes} \code{\link[sf:sfc]{sfc}} object.
}
\description{
This expects \code{graph} to have a \code{lat} and \code{lon} vertex attribute, however you
may pass in alternate numeric vectors to \code{v_lat} and \code{v_lon}.
}
\section{Functions}{
\itemize{
\item \code{edges_to_sf}: Return only the edges as an \code{\link[sf:st_linestring]{st_linestring}} object

\item \code{nodes_to_sf}: Return only the nodes as an \code{\link[sf:st_point]{st_point}} object
}}

