% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{pathway_to_sf}
\alias{pathway_to_sf}
\title{Create an sf object from a konigsberg path}
\usage{
pathway_to_sf(graph, pathway)
}
\arguments{
\item{graph}{A \code{\link{konigsberg_graph}}}

\item{pathway}{A \code{konigsberg_pathway} resulting from \code{\link{cross_all_bridges}}}
}
\value{
A list with two elements:
- \code{pathway} An \code{\link[sf:sf]{sf::sf}} data frame with LINESTRING features
representing the pathway
- \code{terminals} An \code{\link[sf:sf]{sf::sf}} data frame with two POINT features
representing the start and end nodes of the path.
}
\description{
Generates an \code{\link[sf:sf]{sf}} collection of line strings representing the pathway
taken over the map.
}
