% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{way_filters}
\alias{way_filters}
\alias{automobile_highways}
\alias{pedestrian_highways}
\title{OSM Way filters}
\usage{
automobile_highways(graph)

pedestrian_highways(graph)
}
\arguments{
\item{graph}{An \code{\link{konigsberg_graph}} object.}
}
\value{
An integer vector of OSM Way ids.
}
\description{
A variety of presets that will
}
\section{Functions}{
\itemize{
\item \code{automobile_highways}: Ways that are navigable in an automobile

\item \code{pedestrian_highways}: Ways that are navigable by foot
}}

